Code refactoring complete and __init__.py created
