+++ ECommerce_Churn_Propensity_Model/src/config.py
@@ -1,10 +1,5 @@
-import os
-import sys
-import contextlib
-
 class Config:
     def __init__(self):
-        self.model_ver = 0.01
         self.seed = 314
         self.content_file = 'C:/Sample Data/Ecommerce_Churn_Data/ECommerce_Dataset.xlsx'
         self.data_path = 'C:/Sample Data/Ecommerce_Churn_Data'
@@ -14,17 +9,3 @@ class Config:
         self.categorical_cols = ['PreferredLoginDevice', 'CityTier', 'PreferredPaymentMode', 'Gender', 'PreferedOrderCat', 'SatisfactionScore', 'MaritalStatus', 'Complain', 'Churn', 'CouponUsed']
         self.onehot_cols = ['Tenure', 'PreferredLoginDevice', 'PreferredPaymentMode', 'Gender', 'PreferedOrderCat', 'MaritalStatus']
         self.skewed_interval_cols = ['WarehouseToHome', 'Tenure', 'CashbackAmount', 'DaySinceLastOrder', 'OrderCount']
-
-@contextlib.contextmanager
-def suppress_stdout():
-    """
-    For any library that contains (undesirably) verbose output, use this boilerplate function to suppress
-    that output in the CLI.
-    """
-    with open(os.devnull, 'w') as devnull:
-        old_stdout = sys.stdout
-        sys.stdout = devnull
-        try:
-            yield
-        finally:
-            sys.stdout = old_stdout
